Fix script generating PancakeSwap tokens in CI (#1046)
fix script generating PancakeSwap tokens in CI
