+++ package.json
@@ -19,8 +19,8 @@
     "extract-translation-keys": "i18next",
     "generate-subgraph-types": "./scripts/runGenerateSubgraphTypes.sh",
     "generate-contract-types": "rimraf src/types/contracts && typechain --target=ethers-v5 --out-dir=src/types/contracts 'src/constants/contracts/abis/*.json'",
-    "generate-pancake-swap-tokens": "ts-node --esm ./scripts/generatePancakeSwapTokens.mts && yarn prettier --write ./src/constants/tokens/swap/mainnetPancakeSwapTokens.ts",
-    "postinstall": "yarn generate-contract-types && yarn generate-subgraph-types && genversion --es6 --semi src/constants/version.ts",
+    "generate-pancake-swap-tokens": "node ./scripts/generatePancakeSwapTokens.mjs && yarn prettier --write ./src/constants/tokens/swap/mainnetPancakeSwapTokens.ts",
+    "postinstall": "yarn generate-contract-types && yarn generate-subgraph-types && yarn generate-pancake-swap-tokens && genversion --es6 --semi src/constants/version.ts",
     "storybook": "STORYBOOK=true start-storybook -p 6006 -s public",
     "build-storybook": "STORYBOOK=true build-storybook -s public",
     "loki": "loki update --reactUri file:./storybook-static",
